novena-image: Improve real-disk, package list reliability
We now properly write to real disks, and don't fail if no U-Boot is
specified.  Additionally, the package list is now properly parsed.

+++ novena-image.sh
@@ -3,25 +3,7 @@
 rootpass="kosagi"
 version="1.0"
 mirror="http://127.0.0.1:3142/ftp.hk.debian.org/debian"
-packages="sudo openssh-server ntp ntpdate dosfstools btrfs-tools novena-eeprom"
-packages="${packages} xserver-xorg-video-modesetting task-xfce-desktop"
-packages="${packages} hicolor-icon-theme gnome-icon-theme tango-icon-theme"
-packages="${packages} keychain avahi-daemon avahi-dnsconfd libnss-mdns"
-packages="${packages} btrfs-tools dosfstools parted debootstrap python"
-packages="${packages} build-essential xscreensaver vlc vim emacs"
-packages="${packages} x11-xserver-utils usbutils unzip apt-file xz-utils"
-packages="${packages} subversion git make screen tmux read-edid powertop"
-packages="${packages} powermgmt-base pavucontrol p7zip-full paprefs"
-packages="${packages} pciutils nmap ntfs-3g network-manager-vpnc"
-packages="${packages} network-manager-pptp network-manager-openvpn"
-packages="${packages} network-manager-iodine mplayer2 libreoffice"
-packages="${packages} imagemagick icedove iceweasel gtkwave gnupg2 "
-packages="${packages} git git-email git-man fuse freecad enigmail dc"
-packages="${packages} curl clang bridge-utils bluez bluez-tools"
-packages="${packages} bluez-hcidump bison bc automake autoconf pidgin"
-packages="${packages} alsa-utils verilog i2c-tools hwinfo gtkwave"
-packages="${packages} android-tools-adb android-tools-fastboot"
-packages="${packages} android-tools-fsutils"
+packages=""
 debs=""
 
 # Indicates whether we're bootstrapping onto a real disk
@@ -32,6 +14,10 @@ things_mounted=0
 
 info() {
 	func="${FUNCNAME[1]}"
+	if [ -z ${func} ]
+	then
+		func="main"
+	fi
 	FG="1;32m"
 	BG="40m"
 	echo -e "[\033[${FG}\033[${BG}${func}\033[0m] $*"
@@ -39,6 +25,10 @@ info() {
 
 warn() {
 	func="${FUNCNAME[1]}"
+	if [ -z ${func} ]
+	then
+		func="main"
+	fi
 	FG="1;31m"
 	BG="40m"
 	echo -e "[\033[${FG}\033[${BG}${func}\033[0m] $*"
@@ -46,6 +36,10 @@ warn() {
 
 fail() {
 	func="${FUNCNAME[1]}"
+	if [ -z ${func} ]
+	then
+		func="main"
+	fi
 	FG="1;31m"
 	BG="40m"
 	echo -en "[\033[${FG}\033[${BG}${func}\033[0m] "
@@ -58,12 +52,17 @@ fail() {
 	exit 1
 }
 
-cleanup() {
-	info "Unmounting devices from chroot"
-	for mnt in $(grep "${root}" /proc/mounts | awk '{print $2}' | sort -r | uniq)
+unmount_in_dir() {
+	local dir="$1"
+	for mnt in $(grep "${dir}" /proc/mounts | awk '{print $2}' | sort -r | uniq)
 	do
 		umount "${mnt}" 2> /dev/null || warn "Unable to umount ${mnt}"
 	done
+}
+
+cleanup() {
+	info "Unmounting devices from chroot"
+	unmount_in_dir "${root}"
 }
 
 partition_disk() {
@@ -131,8 +130,15 @@ EOF
 prepare_disk() {
 	local diskname="$1"
 	local disktype="$2"
+	local root="$3"
 
 	partition_disk "${diskname}" "${disktype}"
+
+	# MMC devices are weirdly labeled mmcblk0p1 rather than mmcblk01
+	if echo "${diskname}" | grep -q mmcblk
+	then
+		diskname="${diskname}p"
+	fi
 
 	mkfs.vfat ${diskname}1 || fail "Unable to make boot partition"
 	mkswap -f ${diskname}2 || fail "Unable to make swap"
@@ -140,7 +146,7 @@ prepare_disk() {
 
 	mkdir -p "${root}" || fail "Unable to create factory mount directory"
 	mount ${diskname}3 "${root}" || fail "Unable to mount new root filesystem"
-	mkdir "${root}/boot" || fail "Unable to create boot directory"
+	mkdir -p "${root}/boot" || fail "Unable to create boot directory"
 	mount ${diskname}1 "${root}/boot" || fail "Unable to mount new boot filesystem"
 
 }
@@ -163,6 +169,7 @@ prepare_root() {
 	fi
 
 	info "Binding useful mountpoints into new root"
+	mkdir -p "${root}"
 	mount -obind /proc "${root}/proc"
 	mount -obind /sys "${root}/sys"
 	mount -obind /dev "${root}/dev"
@@ -202,8 +209,16 @@ apt_install() {
 	info "Updating package listing"
 	chroot "${root}" apt-get -y update || fail "Couldn't update packages"
 
-	info "Installing basic packages: ${packages}"
-	chroot "${root}" apt-get -y install ${packages} || fail "Couldn't install packages"
+	info "Cleaning up previous apt-get (if any)"
+	chroot "${root}" apt-get -y -f install || fail "Couldn't clean up"
+
+	if [ -z "${packages}" ]
+	then
+		info "No packages were requested to be installed"
+	else
+		info "Installing selected packages: ${packages}"
+		chroot "${root}" apt-get -y install ${packages} || fail "Couldn't install packages"
+	fi
 
 	info "Cleaning up downloaded debs"
 	chroot "${root}" apt-get -y clean || fail "Couldn't clean packages"
@@ -254,6 +269,13 @@ write_uboot_spl() {
 	local spl="$2"
 	local device="$3"
 
+	if [ ! -e "${root}/${spl}" ]
+	then
+		warn "SPL file '${spl}' does not exist, disk won't boot"
+		return 1
+	fi
+
+	info "Writing U-Boot SPL file '${spl}' to disk"
 	dd if="${root}/${spl}" of="${device}" bs=1024 seek=1 conv=notrunc || fail "Unable to write SPL"
 }
 
@@ -273,6 +295,8 @@ usage() {
 	echo "                   you specify a --type as well."
 	echo "    -t  --type     Either 'mmc' or 'sata', the type of disk"
 	echo "                   specified by --disk."
+	echo "    -r  --root     Directory to install files into.  If no --disk"
+	echo "                   is specified, then this argument is required."
 	echo "    -p  --rootpass Which root password to use.  If unspecified,"
 	echo "                   defaults to 'kosagi'."
 	echo "    -l  --packages Specify a comma-separated list of packages"
@@ -291,83 +315,83 @@ usage() {
 
 ##########################################################
 
-main() {
-	temp=`getopt -o m:d:t:p:r:l:s:a:h \
-		--long mirror:,disk:,type:,rootpass:,root:,packages:,suite:,add-deb:,help \
-		-n 'novena-image' -- "$@"`
-	if [ $? != 0 ] ; then fail "Terminating..." >&2 ; exit 1 ; fi
-	eval set -- "$temp"
-	while true ; do
-		case "$1" in
-			-m|--mirror) mirror="$2"; shift 2 ;;
-			-d|--disk) diskname="$2"; shift 2 ;;
-			-t|--type) disktype="$2"; shift 2 ;;
-			-p|--rootpass) rootpass="$2"; shift 2 ;;
-			-r|--root) root="$2"; shift 2 ;;
-			-l|--packages) packages="$2"; shift 2 ;;
-			-s|--suite) suite="$2"; shift 2 ;;
-			-a|--add-deb) debs="${debs} $2"; shift 2 ;;
-			-h|--help) usage; exit 0 ;;
-			--) shift ; break ;;
-			*) fail "Internal getopt error!" ; exit 1 ;;
-		esac
-	done
+temp=`getopt -o m:d:t:p:r:l:s:a:h \
+	--long mirror:,disk:,type:,rootpass:,root:,packages:,suite:,add-deb:,help \
+	-n 'novena-image' -- "$@"`
+if [ $? != 0 ] ; then fail "Terminating..." >&2 ; exit 1 ; fi
+eval set -- "$temp"
+while true ; do
+	case "$1" in
+		-m|--mirror) mirror="$2"; shift 2 ;;
+		-d|--disk) diskname="$2"; shift 2 ;;
+		-t|--type) disktype="$2"; shift 2 ;;
+		-p|--rootpass) rootpass="$2"; shift 2 ;;
+		-r|--root) root="$2"; shift 2 ;;
+		-l|--packages) packages="$2"; info "Packages: $2"; shift 2 ;;
+		-s|--suite) suite="$2"; shift 2 ;;
+		-a|--add-deb) debs="${debs} $2"; shift 2 ;;
+		-h|--help) usage; exit 0 ;;
+		--) shift ; break ;;
+		*) fail "Internal getopt error!" ; exit 1 ;;
+	esac
+done
 
+if [ -z "${suite}" ]
+then
+	fail "Must specify a suite (e.g. jessie) with -s or --suite"
+fi
+
+if [ "$(id -u)" != "0" ]; then
+	fail "As scary as it is, this script must be run as root"
+fi
+
+# Unmount things, and generally clean up on exit
+trap cleanup EXIT
+
+# If a disk path and a type are specified, we're writing to a real disk
+if [ ! -z "${diskname}" ] && [ ! -z "${disktype}" ]
+then
+	realdisk=1
+	info "Ensuring disk is unmounted"
+	unmount_in_dir "${diskname}"
 	if [ -z "${root}" ]
 	then
-		fail "Must specify a root directory with -r or --root"
+		root="/tmp/newroot"
 	fi
+	prepare_disk "${diskname}" "${disktype}" "${root}"
+elif [ ! -z "${diskname}" ] || [ ! -z "${disktype}" ]
+then
+	fail "Must specify both a --disk path and a disk --type"
+elif [ -z "${root}" ]
+then
+	fail "Must specify a root directory with -r or --root"
+fi
 
-	if [ -z "${suite}" ]
-	then
-		fail "Must specify a suite (e.g. jessie) with -s or --suite"
-	fi
+bootstrap "${suite}" "${root}" "${mirror}"
 
-	if [ "$(id -u)" != "0" ]; then
-		fail "As scary as it is, this script must be run as root"
-	fi
+prepare_root "${root}"
 
-	# Unmount things, and generally clean up on exit
-	trap cleanup EXIT
+reset_password "${root}" "${rootpass}"
 
-	# If a disk path and a type are specified, we're writing to a real disk
-	if [ ! -z "${diskname}" ] && [ ! -z "${disktype}" ]
-	then
-		realdisk=1
-		prepare_disk "${diskname}" "${disktype}"
-	elif [ ! -z "${diskname}" ] || [ ! -z "${disktype}" ]
-	then
-		fail "Must specify both a disk path and a disk type"
-	fi
+add_sources "${root}"
 
-	bootstrap "${suite}" "${root}" "${mirror}"
+info "Selected packages: '${packages}'"
+apt_install "${root}" "${packages}"
 
-	prepare_root "${root}"
+if [ ! -z "${debs}" ]
+then
+	deb_install "${root}" ${debs}
+else
+	info "No additional .deb files were requested"
+fi
 
-	reset_password "${root}" "${rootpass}"
+configure_fstab "${root}"
 
-	add_sources "${root}"
+remove_ssh_keys "${root}"
 
-	apt_install "${root}" "${packages}"
+finalize_root "${root}"
 
-	if [ ! -z "${debs}" ]
-	then
-		deb_install "${root}" ${debs}
-	else
-		info "No additional .deb files were requested"
-	fi
-
-	configure_fstab "${root}"
-
-	remove_ssh_keys "${root}"
-
-	finalize_root "${root}"
-
-	if [ ${realdisk} -ne 0 ]
-	then
-		write_uboot_spl "${root}" "/boot/SPL" "${diskname}"
-	fi
-
-}
-
-main $*
+if [ ${realdisk} -ne 0 ]
+then
+	write_uboot_spl "${root}" "/boot/u-boot.spl" "${diskname}"
+fi
